+++ src/black/mode.py
@@ -3,7 +3,6 @@
 Mostly around Python language feature support per version and Black configuration
 chosen by the user.
 """
-
 from hashlib import sha256
 import sys