+++ src/black/output.py
@@ -2,7 +2,6 @@
 
 The double calls are for patching purposes in tests.
 """
-
 import json
 from typing import Any, Optional
 from mypy_extensions import mypyc_attr